Removes playbook hook again.
